+++ packages/plugin-video/web/src/video-component.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { findDOMNode } from 'react-dom';
 import { mergeStyles } from 'wix-rich-content-common';
+import { Loader } from 'wix-rich-content-editor-common';
 import VideoViewer from './video-viewer';
 import styles from '../statics/styles/default-video-styles.scss';
 import { VIDEO_TYPE_LEGACY, VIDEO_TYPE } from './types';
@@ -21,7 +22,6 @@ class VideoComponent extends React.Component {
     super(props);
     const isPlayable = !props.blockProps;
     this.state = {
-      isLoading: false,
       isLoaded: false,
       isPlayable,
     };
@@ -48,7 +48,7 @@ class VideoComponent extends React.Component {
   }
 
   handleReady = () => {
-    if (!this.state.isLoaded) {
+    if (!this.state.isLoaded && !this.props.componentData.tempData) {
       this.setState({ isLoaded: true });
     }
   };
@@ -61,6 +61,18 @@ class VideoComponent extends React.Component {
         {isLoaded && <span className={styles.video_overlay_message}>{overlayText}</span>}
       </div>
     );
+  };
+
+  renderLoader = () => {
+    return (
+      <div className={this.styles.videoOverlay}>
+        <Loader type={'medium'} />
+      </div>
+    );
+  };
+
+  onReload = () => {
+    this.setState({ isLoaded: false });
   };
 
   renderPlayer = () => {
@@ -82,6 +94,8 @@ class VideoComponent extends React.Component {
         disableRightClick={disableRightClick}
         theme={theme}
         setComponentUrl={setComponentUrl}
+        onReload={this.onReload}
+        isLoaded={this.state.isLoaded}
       />
     );
   };
@@ -108,6 +122,7 @@ class VideoComponent extends React.Component {
       >
         {!isPlayable && this.renderOverlay(this.styles, this.props.t)}
         {this.renderPlayer()}
+        {!this.state.isLoaded && this.renderLoader()}
       </div>
     );
     /* eslint-enable jsx-a11y/no-static-element-interactions */